// quizonline-frontend/src/app/api/generated/model/enrollment-mode-enum.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `open` - Open enrollment * `approval` - Requires approval * `invite` - Invite-only
 */
export enum EnrollmentModeEnumDto {

    Open = 'open',

    Approval = 'approval',

    Invite = 'invite'
}