// quizonline-frontend/src/app/api/generated/model/level-enum.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `beginner` - Beginner * `intermediate` - Intermediate * `advanced` - Advanced
 */
export enum LevelEnumDto {

    Beginner = 'beginner',

    Intermediate = 'intermediate',

    Advanced = 'advanced'
}